requirements.txt
fastapi
uvicorn
python-dotenv
aiofiles
requests
openai
python-multipart